requirements.txt
psycopg2-binary==2.9.9
pandas==2.0.3
python-dotenv==1.0.0
numpy==1.24.3
schedule==1.2.2